requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
configobj==5.0.8
configparser==7.0.0
cryptography==41.0.1
dataclasses==0.6
defusedxml==0.8.0rc2
Deprecated==1.2.14
dj-database-url==2.2.0
Django==4.2.13
django-background-tasks==1.2.0
django-compat==1.0.15
django-debug-toolbar==2.1
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
et-xmlfile==1.1.0
etelemetry==0.3.1
exceptiongroup==1.2.1
ez-setup==0.9
filelock==3.15.4
gunicorn==22.0.0
h11==0.14.0
httplib2==0.22.0
idna==2.8
isodate==0.6.1
kombu==5.3.7
looseversion==1.3.0
lxml==5.2.2
Markdown==3.1.1
mysqlclient==2.0.3
networkx==3.2.1
nibabel==5.2.1
nipype==1.8.6
numpy==1.25.0
oauthlib==3.2.2
opencv-python==4.8.0.74
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.0.3
pathlib==1.0.1
pikepdf==9.1.0
pillow==10.3.0
pillow_heif==0.16.0
prompt_toolkit==3.0.45
prov==2.0.1
psycopg2-binary==2.9.9
pycparser==2.20
pydot==3.0.1
Pygments==2.4.2
PyJWT==2.8.0
PyMuPDF==1.24.8
PyMuPDFb==1.24.8
pyotp==2.9.0
pyparsing==3.1.2
PyPDF2==3.0.1
pypng==0.20220715.0
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2020.1
pyxnat==1.6.2
qrcode==7.4.2
rdflib==6.3.2
redis==5.0.4
requests==2.32.3
requests-oauthlib==2.0.0
schedule==1.2.2
scipy==1.13.1
selenium==4.22.0
simplejson==3.19.2
six==1.15.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sortedcontainers==2.4.0
sqlparse==0.3.1
traits==6.3.2
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.1
tzdata==2023.3
urllib3==2.2.2
user-agent==0.1.9
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.1
websocket-client==1.8.0
whitenoise==6.6.0
wrapt==1.16.0
wsproto==1.2.0
